requirements.txt
mujoco-warp==3.6.0
mediapy==1.2.6
torch
tensorboard
warp-lang